:hammer: refactor server test structure
