+++ package.json
@@ -17,11 +17,19 @@
     "start:client": "node client/scripts/start.js",
     "build:server": "cross-env NODE_ENV=development babel server/src -d server/build",
     "build:client": "node client/scripts/build.js",
+    "migrate": "sequelize db:migrate:undo:all && sequelize db:migrate",
+    "migrate:test": "sequelize db:migrate:undo:all --env=test && sequelize db:migrate --env=test",
+    "seed": "sequelize db:seed:all",
+    "seed:test": "sequelize db:seed:all --env=test",
     "test:server": "cross-env NODE_ENV=test SECRET=mytestsecret nyc mocha server/tests/test.js --recursive --timeout 30000",
-    "test:server:watch": "cross-env NODE_ENV=test SECRET=mytestsecret nyc mocha server/tests/test.js --recursive --timeout 30000 --watch",
+    "test:server:watch": "cross-env NODE_ENV=test SECRET=mytestsecret mocha server/tests/test.js --recursive --timeout 30000 --watch",
     "test:client": "node client/scripts/test.js --env=jsdom --coverage",
     "test:client:watch": "node client/scripts/test.js --env=jsdom",
     "pretest": "npm run lint",
+    "pretest:server": "npm-run-all migrate:test seed:test",
+    "pretest:server:watch": "npm-run-all migrate:test seed:test",
+    "pretest:client": "npm run lint",
+    "posttest": "lcov-result-merger ./*/coverage/lcov.info",
     "prebuild:init": "cd server && mkdir build",
     "prebuild": "npm-run-all clean prebuild:init",
     "clean": "npm-run-all clean:server clean:client",
@@ -29,13 +37,12 @@
     "clean:client": "cd client && rm -rf build",
     "lint": "eslint **/*.js --fix",
     "docs": "aglio -i docs.apib --theme-variables flatly --theme-template triple -o docs/index.html",
-    "coverage": "nyc report --reporter=text-lcov | coveralls",
+    "coverage": "lcov-result-merger ./*/coverage/lcov.info | coveralls",
     "analyze": "source-map-explorer client/build/static/js/main.*"
   },
   "dependencies": {
     "autoprefixer": "^7.2.2",
     "axios": "^0.17.1",
-    "axios-mock-adapter": "^1.11.0",
     "babel-cli": "^6.26.0",
     "babel-core": "6.26.0",
     "babel-eslint": "^8.0.3",
@@ -78,13 +85,13 @@
     "html-webpack-plugin": "^2.30.1",
     "jquery": "^3.2.1",
     "jsonwebtoken": "^8.1.0",
+    "lcov-result-merger": "^2.0.0",
     "loaders.css": "^0.1.2",
     "lodash": "^4.17.4",
     "minimatch": "3.0.2",
     "mocha": "^3.5.3",
     "morgan": "^1.9.0",
     "moxios": "^0.4.0",
-    "multer": "^1.3.0",
     "nightwatch": "^0.9.19",
     "node-sass": "^4.7.2",
     "normalizr": "^3.2.4",
@@ -113,7 +120,6 @@
     "react-transition-group": "^2.2.1",
     "reactstrap": "^5.0.0-alpha.4",
     "redux": "^3.7.2",
-    "redux-form": "^7.2.0",
     "redux-logger": "^3.0.6",
     "redux-mock-store": "^1.4.0",
     "redux-thunk": "^2.2.0",
@@ -135,7 +141,6 @@
     "yup": "^0.23.0"
   },
   "devDependencies": {
-    "chai-fs": "^2.0.0",
     "coveralls": "^2.13.3",
     "cross-env": "^5.1.1",
     "eslint": "^4.9.0",
@@ -154,7 +159,6 @@
     "jest": "^21.2.1",
     "jest-enzyme": "^4.0.1",
     "mocha-lcov-reporter": "^1.3.0",
-    "node-sass-chokidar": "^0.0.3",
     "nodemon": "^1.12.1",
     "nyc": "^11.3.0"
   }
