--- server/tests/apiroutes/api.test.js
@@ -1,1 +1,0 @@
-import './search.recipe.api.test';
