+++ server/tests/controllers/favorites/index.js
@@ -1,0 +1,2 @@
+import './getFavorite.test';
+import './addFavorite.test';
