+++ server/tests/controllers/index.js
@@ -1,0 +1,6 @@
+import './user';
+import './recipes';
+import './reviews';
+import './votes';
+import './favorites';
+import './search';
