+++ server/tests/controllers/recipes/index.js
@@ -1,0 +1,5 @@
+import './addRecipe.test';
+import './updateRecipe.test';
+import './deleteRecipe.test';
+import './getAllRecipes.test';
+import './getRecipeDetails.test';
