+++ server/tests/controllers/reviews/index.js
@@ -1,0 +1,1 @@
+import './reviews.test';
