+++ server/tests/controllers/search/index.js
@@ -1,0 +1,1 @@
+import './search.test';
