+++ server/tests/controllers/user/index.js
@@ -1,0 +1,4 @@
+import './signup.test';
+import './login.test';
+import './retrieve.test';
+import './refreshToken.test';
