+++ server/tests/controllers/votes/index.js
@@ -1,0 +1,3 @@
+import './getUpvotedRecipes.test';
+import './upvoteRecipe.test';
+import './downvoteRecipe.test';
