+++ server/tests/helpers/arrayCheck.test.js
