--- server/tests/helpers/helpers.test.js
@@ -1,3 +1,0 @@
-import './passwordhash.helpers.test';
-import './arrayCheck.helpers.test';
-import './getCleanUser.helpers.test';
