+++ server/tests/helpers/index.js
@@ -1,0 +1,3 @@
+import './passwordhash.test';
+import './arrayCheck.test';
+import './getCleanUser.test';
