+++ server/tests/home.test.js
@@ -28,4 +28,42 @@ describe('App Home', () => {
         });
     });
   });
+
+  describe('API Route', () => {
+    it('should return a Welcome Message for API Home', (done) => {
+      request(app)
+        .get('/api')
+        .end((err, res) => {
+          expect(res.statusCode).to.equal(200);
+          expect(res.body.message).to.equal('Welcome to the More Recipes API');
+
+          if (err) return done(err);
+          done();
+        });
+    });
+
+    it('should return a Welcome Message for Version 1 API Home', (done) => {
+      request(app)
+        .get('/api/v1')
+        .end((err, res) => {
+          expect(res.statusCode).to.equal(200);
+          expect(res.body.message).to.equal('Welcome to version 1 of the More Recipes API');
+
+          if (err) return done(err);
+          done();
+        });
+    });
+
+    it('should return a Fallback Message for API Route', (done) => {
+      request(app)
+        .get('/api/fallback')
+        .end((err, res) => {
+          expect(res.statusCode).to.equal(409);
+          expect(res.body.message).to.equal('Where Are You Going? Page Not Found');
+
+          if (err) return done(err);
+          done();
+        });
+    });
+  });
 });
