+++ server/tests/middlewares/errorHandler.test.js
