+++ server/tests/middlewares/index.js
@@ -1,0 +1,1 @@
+import './errorHandler.test';
