--- server/tests/middlewares/middlewares.test.js
@@ -1,1 +1,0 @@
-import './errorHandler.middlewares.test';
